requirements.txt
numpy==1.26.4
pillow==10.3.0
wxPython==4.2.1
